apps/analytics-service/pyproject.toml
[project]
name = "sentinel-analytics"
version = "0.1.0"
description = "SENTINEL Analytics Service - ML and pattern analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "confluent-kafka>=2.4.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "shapely>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0", "pytest-asyncio>=0.23.0"]
